add tox.ini and .travis.yml files
To start using Travis-CI, we just need to log into Travis, and set up
the webhook with Github. Directions are here:
https://docs.travis-ci.com/user/getting-started/

The travis file tells TravisCI what to do for the tests. In this case, I
set it up to run tox, which uses tox.ini. Tox will run all of the tests
in python 2.7, 3.3, 3.4, and 3.5. We can also configure it to test the
documentation, do PEP8 checking (not worth it, IMHO), or even to build
the documentation with Sphinx, or deploy to PyPI. For now, we'll just do
our testing!